+++ app/Posts.php
@@ -8,4 +8,7 @@ class Posts extends Model {
     public function upload() {
         return $this->hasOne(Uploads::class, "post_id");
     }
+    public function getContentAttribute($content) {
+        return nl2br($content);
+    }
 }